+++ ftxui/src/ftxui/component/menu.cpp
@@ -8,7 +8,7 @@ Element Menu::Render() {
   std::vector<Element> elements;
   bool focused = Focused();
   for (size_t i = 0; i < entries.size(); ++i) {
-    if (size_t(selected) == i) {
+    if (selected == i) {
       if (focused)
         elements.push_back(active_style(text(L"> " + entries[i])));
       else
@@ -25,12 +25,12 @@ bool Menu::OnEvent(Event event) {
   if (!Focused())
     return false;
 
-  int new_selected = selected;
+  size_t new_selected = selected;
   if (event == Event::ArrowUp || event == Event::Character('k'))
     new_selected--;
   if (event == Event::ArrowDown || event == Event::Character('j'))
     new_selected++;
-  new_selected = std::max(0, std::min(int(entries.size())-1, new_selected));
+  new_selected = std::max(size_t(0), std::min(entries.size()-size_t(1), new_selected));
 
   if (selected != new_selected) {
     selected = new_selected;
